rejoined BigInteger file after semplifications
