+++ test/Deveel.Math.Core.Tests/bin/Debug/netcoreapp1.1/Deveel.Math.Core.Tests.deps.json
@@ -9,7 +9,7 @@
       "deveel.math.core.tests/1.0.0": {
         "dependencies": {
           "Microsoft.NET.Test.Sdk": "15.0.0",
-          "dmath.core": "1.0.0",
+          "dmath.core": "1.6.1",
           "xunit": "2.2.0",
           "xunit.runner.visualstudio": "2.2.0"
         },
@@ -400,7 +400,7 @@
           "Microsoft.TestPlatform.ObjectModel": "15.0.0"
         }
       },
-      "dmath.core/1.0.0": {
+      "dmath.core/1.6.1": {
         "runtime": {
           "Deveel.Math.Core.dll": {}
         }
@@ -588,7 +588,7 @@
       "path": "xunit.runner.visualstudio/2.2.0",
       "hashPath": "xunit.runner.visualstudio.2.2.0.nupkg.sha512"
     },
-    "dmath.core/1.0.0": {
+    "dmath.core/1.6.1": {
       "type": "project",
       "serviceable": false,
       "sha512": ""
